editoast: rename `mod auth` to `mod auth_driver`
