+++ editoast/src/views/test_app.rs
@@ -32,7 +32,7 @@ use crate::core::CoreClient;
 use crate::generated_data::speed_limit_tags_config::SpeedLimitTagIds;
 use crate::infra_cache::InfraCache;
 use crate::map::MapLayers;
-use crate::models::auth::PgAuthDriver;
+use crate::models::PgAuthDriver;
 use crate::valkey_utils::ValkeyConfig;
 use crate::AppState;
 use crate::ValkeyClient;
